requirements.txt
PyYAML>=6.0
numpy>=1.24
pandas>=2.0
yfinance>=0.2.40
gymnasium>=0.29
ta>=0.11.0
stable-baselines3>=2.3.0
tqdm>=4.66
tensorboard>=2.16
